+++ deal.II/include/deal.II/dofs/block_info.h
@@ -261,7 +261,7 @@ BlockInfo::level (const unsigned int l) const
 inline
 types::global_dof_index BlockInfo::renumber (const unsigned int i) const
 {
-  AssertIndexRange(i, local_renumbering.size());
+  AssertIndexRange(i, static_cast<unsigned int>(local_renumbering.size()));
   return local_renumbering[i];
 }
